+++ src/main/java/org/folio/cql2rmapi/CQLParserForRMAPI.java
@@ -135,8 +135,12 @@ public class CQLParserForRMAPI {
       setSelection(termNode);
       break;
     default:
-      if(!Stream.of(TITLE, CODEX_TITLE, IDENTIFIER, CODEX_IDENTIFIER, PUBLISHER, CODEX_PUBLISHER, ID, CODEX_ID).anyMatch(indexNode::equalsIgnoreCase)) {
-     // If search field is not supported, log and return an error response
+      if (indexNode.startsWith("ext.")) {
+        // CQL fields that are in the ext context set should be ignored if they
+        // are not recognized by the module.
+        break;
+      } else if(!Stream.of(TITLE, CODEX_TITLE, IDENTIFIER, CODEX_IDENTIFIER, PUBLISHER, CODEX_PUBLISHER, ID, CODEX_ID).anyMatch(indexNode::equalsIgnoreCase)) {
+        // If search field is not supported, log and return an error response
         builder.append("Search field or filter value ");
         builder.append(indexNode);
         builder.append(UNSUPPORTED);
@@ -274,9 +278,9 @@ public class CQLParserForRMAPI {
         builder.append("&resourcetype=" + filterValue);
       }
 
+      final String queryParam = "&selection=";
       if (selection != null) {
         // Map fields to RM API
-        final String queryParam = "&selection=";
         switch(selection.toLowerCase()) {
         case "all":
           builder.append(queryParam);
@@ -298,6 +302,11 @@ public class CQLParserForRMAPI {
           errorMsgBuilder.append(UNSUPPORTED);
           throw new QueryValidationException(errorMsgBuilder.toString());
         }
+      } else {
+        // Default "selection" to "selected" if the CQL did not contain the
+        // "ext.selected" field.
+        builder.append(queryParam);
+        builder.append("selected");
       }
 
       builder.append("&orderby=" + sortType);