+++ src/test/java/org/folio/cql2rmapi/CQLParserForRMAPITest.java
@@ -1,12 +1,17 @@
 package org.folio.cql2rmapi;
 
 import static org.hamcrest.CoreMatchers.instanceOf;
+import static org.hamcrest.Matchers.containsString;
+import static org.hamcrest.Matchers.stringContainsInOrder;
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
 
 import java.io.UnsupportedEncodingException;
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 
 import org.junit.Test;
 import org.z3950.zing.cql.CQLNode;
@@ -126,7 +131,7 @@ public class CQLParserForRMAPITest {
   public void CQLParserReturnsExpectedQueryStringIfValidQueryIsPassedTest()
       throws QueryValidationException, UnsupportedEncodingException {
     final CQLParserForRMAPI parser = new CQLParserForRMAPI(VALID_QUERY, 10, 10);
-    assertEquals("search=bridget&searchfield=titlename&orderby=titlename&count=10&offset=2", parser.getRMAPIQueries().get(0));
+    assertEquals("search=bridget&searchfield=titlename&selection=selected&orderby=titlename&count=10&offset=2", parser.getRMAPIQueries().get(0));
   }
 
   @Test(expected = QueryValidationException.class)
@@ -154,39 +159,39 @@ public class CQLParserForRMAPITest {
   @Test
   public void CQLParserMapsIdentifierToISXNTest() throws QueryValidationException, UnsupportedEncodingException {
     final CQLParserForRMAPI parser = new CQLParserForRMAPI(VALID_ISBN_QUERY, 0, 10);
-    assertEquals("search=12345&searchfield=isxn&orderby=titlename&count=10&offset=1", parser.getRMAPIQueries().get(0));
+    assertEquals("search=12345&searchfield=isxn&selection=selected&orderby=titlename&count=10&offset=1", parser.getRMAPIQueries().get(0));
   }
 
   @Test
   public void CQLParserPassesAlongOffsetIfNotEqualToZeroTest() throws QueryValidationException, UnsupportedEncodingException {
     final CQLParserForRMAPI parser = new CQLParserForRMAPI(VALID_ISBN_QUERY, 10, 10);
-    assertEquals("search=12345&searchfield=isxn&orderby=titlename&count=10&offset=2", parser.getRMAPIQueries().get(0));
+    assertEquals("search=12345&searchfield=isxn&selection=selected&orderby=titlename&count=10&offset=2", parser.getRMAPIQueries().get(0));
   }
 
   @Test
   public void CQLParserDoesNotAppendResourceTypeIfTypeNotPassedAlongTest() throws QueryValidationException, UnsupportedEncodingException {
     final CQLParserForRMAPI parser = new CQLParserForRMAPI(VALID_ISBN_QUERY, 10, 10);
-    assertEquals("search=12345&searchfield=isxn&orderby=titlename&count=10&offset=2", parser.getRMAPIQueries().get(0));
+    assertEquals("search=12345&searchfield=isxn&selection=selected&orderby=titlename&count=10&offset=2", parser.getRMAPIQueries().get(0));
   }
 
   @Test
   public void CQLParserDoesNotAppendResourceTypeIfFilterTypeIsNullTest() throws QueryValidationException, UnsupportedEncodingException {
     final CQLParserForRMAPI parser = new CQLParserForRMAPI(VALID_FILTER_QUERY, 10, 10);
     parser.filterType = null;
-    assertEquals("search=bridget&searchfield=titlename&orderby=titlename&count=10&offset=1", parser.buildRMAPIQuery(10, 1));
+    assertEquals("search=bridget&searchfield=titlename&selection=selected&orderby=titlename&count=10&offset=1", parser.buildRMAPIQuery(10, 1));
   }
 
   @Test
   public void CQLParserDoesNotAppendResourceTypeIfFilTerValueIsNullTest() throws QueryValidationException, UnsupportedEncodingException {
     final CQLParserForRMAPI parser = new CQLParserForRMAPI(VALID_FILTER_QUERY, 10, 10);
     parser.filterValue = null;
-    assertEquals("search=bridget&searchfield=titlename&orderby=titlename&count=10&offset=1", parser.buildRMAPIQuery(10, 1));
+    assertEquals("search=bridget&searchfield=titlename&selection=selected&orderby=titlename&count=10&offset=1", parser.buildRMAPIQuery(10, 1));
   }
 
   @Test
   public void CQLParserAppendsResourceTypeIfFilterTypeAndFilTerValueAreValidTest() throws QueryValidationException, UnsupportedEncodingException {
     final CQLParserForRMAPI parser = new CQLParserForRMAPI(VALID_FILTER_QUERY, 10, 10);
-    assertEquals("search=bridget&searchfield=titlename&resourcetype=streamingvideo&orderby=titlename&count=10&offset=2",
+    assertEquals("search=bridget&searchfield=titlename&resourcetype=streamingvideo&selection=selected&orderby=titlename&count=10&offset=2",
         parser.getRMAPIQueries().get(0));
   }
 
@@ -225,7 +230,7 @@ public class CQLParserForRMAPITest {
     assertEquals(1, queries.size());
     assertEquals(0, parser.getInstanceIndex());
     for (final String query: queries) {
-      assertEquals("search=bridget&searchfield=titlename&orderby=titlename&count=100&offset=10", query);
+      assertEquals("search=bridget&searchfield=titlename&selection=selected&orderby=titlename&count=100&offset=10", query);
     }
   }
 
@@ -235,8 +240,8 @@ public class CQLParserForRMAPITest {
     final ArrayList<String> queries = (ArrayList<String>) parser.getRMAPIQueries();
     assertEquals(2, queries.size());
     assertEquals(1, parser.getInstanceIndex());
-    assertEquals("search=bridget&searchfield=titlename&orderby=titlename&count=15&offset=3", queries.get(0));
-    assertEquals("search=bridget&searchfield=titlename&orderby=titlename&count=15&offset=4", queries.get(1));
+    assertEquals("search=bridget&searchfield=titlename&selection=selected&orderby=titlename&count=15&offset=3", queries.get(0));
+    assertEquals("search=bridget&searchfield=titlename&selection=selected&orderby=titlename&count=15&offset=4", queries.get(1));
   }
 
   @Test(expected = QueryValidationException.class)
@@ -360,5 +365,25 @@ public class CQLParserForRMAPITest {
     final CQLParserForRMAPI parser = new CQLParserForRMAPI("id=12345" , 0, 1);
     assertTrue(parser.isIDSearchField());
     assertEquals("12345", parser.getID());
+  }
+
+  @Test
+  public void CQLParserUnknownExt() throws Exception {
+    final CQLParserForRMAPI parser = new CQLParserForRMAPI("title=xyzzy and ext.available=all", 0, 10);
+    final List<String> queries = parser.getRMAPIQueries();
+    assertEquals(1, queries.size());
+    String query = queries.get(0);
+    assertNotNull(query);
+    assertThat(query, stringContainsInOrder(Arrays.asList("search=xyzzy", "searchfield=title")));
+  }
+
+  @Test
+  public void CQLParserSelectedNotPresent() throws Exception {
+    final CQLParserForRMAPI parser = new CQLParserForRMAPI("title=xyzzy", 0, 10);
+    final List<String> queries = parser.getRMAPIQueries();
+    assertEquals(1, queries.size());
+    String query = queries.get(0);
+    assertNotNull(query);
+    assertThat(query, containsString("selection=selected"));
   }
 }